Move OLED code directly into keymap.c
